mailing-list: Only flush orm manager when there are new subscribers

+++ src/MailingListBundle/SubscriberManager.php
@@ -33,6 +33,10 @@ class SubscriberManager
 
     public function flush()
     {
+        if (empty($this->signups)) {
+            return;
+        }
+
         foreach ($this->signups as $subscriber) {
             $this->connector->subscribe($subscriber);
             $this->em->remove($subscriber);
